Refactoring: Whole package was rewritten as static class for purpose of better code organization.
