+++ packages/mockup-schema/assets/generated-variables.txt
