+++ packages/mockup-schema/source/helpers.d.ts
@@ -1,0 +1,1 @@
+declare type VariableRecord<T = string> = Record<string, T | object>;
