+++ packages/mockup-schema/style-schema.json
